requirements.txt
playwright
playwright-stealth
httpx
python-dotenv
python-dateutil
tenacity
pytest
pytest-mock
pytest-asyncio
